+++ bakedanuki/bakedanuki-util/python/bd_util/maya/node/__init__.py
@@ -1,16 +1,22 @@
 # coding: utf-8
 from .nodes import Nodes
 from .inspection import (
+    ScalarAttributeDisplayFilter,
     ScalarAttributeInfo,
     ScalarAttributeKind,
+    filter_scalar_attribute_paths,
     inspect_scalar_attributes,
+    matches_scalar_attribute_display_filter,
     selected_node_names,
 )
 
 __all__ = (
     "Nodes",
+    "ScalarAttributeDisplayFilter",
     "ScalarAttributeInfo",
     "ScalarAttributeKind",
+    "filter_scalar_attribute_paths",
     "inspect_scalar_attributes",
+    "matches_scalar_attribute_display_filter",
     "selected_node_names",
 )